разделил и структурировал вьюшки и контролер на несколько файлов
