+++ popitka/Views/Account/UserProfile.cshtml
@@ -6,9 +6,9 @@
     <header class="bg-white shadow-sm sticky top-0 z-50">
 
         <div id="profileDropdown" class="hidden absolute right-0 mt-2 w-48 bg-white rounded-md shadow-lg py-1">
-            <a href="/Home/Profile" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">Профиль</a>
-            <a href="/Home/Order" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">История заказов</a>
-            <a href="/Home/Login" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">Выйти</a>
+            <a href="/Account/Profile" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">Профиль</a>
+            <a href="/Order/Order" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">История заказов</a>
+            <a href="/Account/Login" class="block px-4 py-2 text-sm text-gray-700 hover:bg-gray-100">Выйти</a>
         </div>
     </header>
 
@@ -22,7 +22,7 @@
                     <div class="bg-white rounded-lg shadow-md p-6">
                         <nav class="space-y-2">
                             <a href="#profile" class="block px-4 py-2 rounded-md bg-blue-50 text-blue-600">Профиль</a>
-                            <a href="/Home/Order" class="block px-4 py-2 rounded-md text-gray-700 hover:bg-gray-50">История заказов</a>
+                            <a href="/Order/Order" class="block px-4 py-2 rounded-md text-gray-700 hover:bg-gray-50">История заказов</a>
                             <a href="#settings" class="block px-4 py-2 rounded-md text-gray-700 hover:bg-gray-50">Настройки</a>
                         </nav>
                     </div>
