+++ popitka/Views/Cart/Cart.cshtml
@@ -40,7 +40,7 @@
         </p>
     }
     <div class="flex justify-end">
-        <a href="@Url.Action("Checkout", "Home")" class="bg-blue-600 text-white px-6 py-2 rounded-lg hover:bg-blue-700 transition inline-block">
+        <a href="@Url.Action("Checkout", "Order")" class="bg-blue-600 text-white px-6 py-2 rounded-lg hover:bg-blue-700 transition inline-block">
             Оформить заказ
         </a>
     </div>
