+++ popitka/Views/Product/Index.cshtml
@@ -32,24 +32,24 @@
                 @foreach (var product in Model)
                 {
                     <div class="product-card bg-white rounded-lg shadow-md p-4 text-center hover:shadow-lg transition">
-                        <a href="@Url.Action("ProductDetails", "Home", new { id = product.Id })" style="text-decoration: none; color: inherit;">
+                        <a href="@Url.Action("ProductDetails", "Product", new { id = product.Id })" style="text-decoration: none; color: inherit;">
                             <img src="@product.ImageUrl" alt="@product.Name" class="product-image mb-4 w-full h-48 object-contain" />
                             <h3 class="text-xl font-semibold">@product.Name</h3>
                             <p class="text-gray-600 mb-2">@product.Description</p>
                             <p class="text-blue-600 font-bold mb-4">@product.Price.ToString("C")</p>
                         </a>
 
-                        <a href="@Url.Action("AddToCart", "Home", new { productId = product.Id })" class="btn btn-sm bg-blue-500 text-white rounded px-3 py-1 mb-2 inline-block">
+                        <a href="@Url.Action("AddToCart", "Cart", new { productId = product.Id })" class="btn btn-sm bg-blue-500 text-white rounded px-3 py-1 mb-2 inline-block">
                             В корзину
                         </a>
 
                         @if (Session["IsAdmin"] != null && (bool)Session["IsAdmin"])
                         {
                             <div class="flex justify-center gap-4 mt-2">
-                                <a href="@Url.Action("AddOrEditProduct", "Home", new { id = product.Id })" title="Редактировать" class="text-yellow-500 hover:text-yellow-700 text-xl">
+                                <a href="@Url.Action("AddOrEditProduct", "Product", new { id = product.Id })" title="Редактировать" class="text-yellow-500 hover:text-yellow-700 text-xl">
                                     ✏️
                                 </a>
-                                @using (Html.BeginForm("DeleteProduct", "Home", FormMethod.Post, new { @class = "inline" }))
+                                @using (Html.BeginForm("Index", "Product", FormMethod.Post, new { @class = "inline" }))
                                 {
                                     @Html.Hidden("id", product.Id)
                                     <button type="submit" title="Удалить" onclick="return confirm('Удалить товар?')" class="inline bg-transparent border-none text-red-600">
@@ -77,7 +77,7 @@
 
         if (!confirm("Удалить этот товар?")) return;
 
-        $.post("/Home/DeleteProduct", { id: id })
+        $.post("/Product/Index", { id: id })
             .done(function () {
                 location.reload();
             })
