+++ popitka/Views/Shared/LayoutFooter.cshtml
@@ -5,15 +5,15 @@
             <div>
                 <h3 class="text-lg font-semibold mb-4">О компании</h3>
                 <ul class="space-y-2">
-                    <li><a href="/Home/About" class="text-gray-600 hover:text-blue-600">О нас</a></li>
-                    <li><a href="/Home/Support" class="text-gray-600 hover:text-blue-600">Служба поддержка</a></li>
+                    <li><a href="/Page/About" class="text-gray-600 hover:text-blue-600">О нас</a></li>
+                    <li><a href="/Page/Support" class="text-gray-600 hover:text-blue-600">Служба поддержка</a></li>
                     <li><a href="/careers" class="text-gray-600 hover:text-blue-600">Вакансии</a></li>
                 </ul>
             </div>
             <div>
                 <h3 class="text-lg font-semibold mb-4">Покупателям</h3>
                 <ul class="space-y-2">
-                    <li><a href="/Home/Shipping" class="text-gray-600 hover:text-blue-600">Доставка</a></li>
+                    <li><a href="/Page/Shipping" class="text-gray-600 hover:text-blue-600">Доставка</a></li>
                     <li><a href="/payment" class="text-gray-600 hover:text-blue-600">Оплата</a></li>
                     <li><a href="/return" class="text-gray-600 hover:text-blue-600">Возврат</a></li>
                 </ul>
